server/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
decorator==5.0.7
defusedxml==0.6.0
Django==3.1.7
django-bootstrap-v5==1.0.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.12.4
entrypoints==0.3
Faker==8.1.2
idna==2.10
importlib-metadata==2.1.1
ipykernel==5.4.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.0
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.20.1
packaging==20.8
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==21.0.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
